+++ src/Smdn.Net.EchonetLite/Smdn.Net.EchonetLite/EchonetPropertyExtentions.cs
@@ -3,22 +3,10 @@
 // SPDX-License-Identifier: MIT
 using System.Text;
 
-namespace Smdn.Net.EchonetLite.Extensions
+namespace Smdn.Net.EchonetLite
 {
-    public static class Extentions
+    internal static class EchonetPropertyExtentions
     {
-        public static string GetDebugString(this EchonetObject obj)
-        {
-            if (obj == null)
-            {
-                return "null";
-            }
-            if(obj.Spec == null)
-            {
-                return "Spec null";
-            }
-            return $"0x{obj.Spec.ClassGroup.Code:X2}{obj.Spec.ClassGroup.Name} 0x{obj.Spec.Class.Code:X2}{obj.Spec.Class.Name} {obj.InstanceCode:X2}";
-        }
         public static string GetDebugString(this EchonetProperty property)
         {
             if (property == null)
